DDA implementation, memory issue fixes, test script, new maps+textures

+++ sources/drawing/utils.c
@@ -6,7 +6,7 @@
 /*   By: mekundur <mekundur@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/02/19 16:05:39 by drongier          #+#    #+#             */
-/*   Updated: 2025/04/26 15:16:37 by mekundur         ###   ########.fr       */
+/*   Updated: 2025/05/05 12:02:57 by mekundur         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -22,9 +22,8 @@ bool	touch(float px, float py, t_game *game)
 	char	**map;
 
 	map = game->map->map;
-	x = (int)px / BLOCK;
-	y = (int)py / BLOCK;
-
+	x = px / BLOCK;
+	y = py / BLOCK;
 	if (map[y][x] == '1')
 		return (true);
 	return (false);
